First round of code fixes from manual testing of Postgres functions. Mostly unconverted nextval syntax, a few null-vs-empty-string confusions, places where Postgres is pickier about datatype matches, etc etc.

+++ schema/spacewalk/postgres/procs/sequence_currval.sql
@@ -1,5 +1,5 @@
-create or replace function sequence_currval( seq_name regclass ) returns numeric as
+create or replace function
+sequence_currval( seq_name regclass ) returns bigint as
 $$
-	select currval( $1 )::numeric;
+	select currval($1);
 $$ language sql;
-
